refactor: 데이터 처리 타입 변경
- int > long

+++ src/main/java/calculator/Application.java
@@ -9,7 +9,6 @@ import calculator.utils.IOUtils;
 public class Application {
     public static void main(String[] args) {
         // TODO: 프로그램 구현
-        //I/O 분리: 입력 안내, 결과 출력 형식 맞추기
         try {
             String input = IOUtils.readLine();
 
@@ -20,10 +19,9 @@ public class Application {
                     new SumServiceImpl()
             );
 
-            int result = calculator.add(input);
+            long result = calculator.add(input);
 
-            //I/O 분리: 입력 안내, 결과 출력 형식 맞추기
-            IOUtils.printResult(result);
+            IOUtils.printResult((int) result);
         } catch (IllegalArgumentException e) {
             System.out.println(e.getMessage());
             return;
